+++ handler/root_handler.go
@@ -1,0 +1,8 @@
+package handler
+
+import "net/http"
+
+// Root handler
+func RootHandler(w http.ResponseWriter, r *http.Request) {
+	w.Write([]byte("Hi!"))
+}
